requirements.txt
appdirs==1.4.0
bleach==1.4
catfish==1.4.2
chardet==2.3.0
cupshelpers==1.0
decorator==4.0.10
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.6
django-guardian==1.4.5
gunicorn==19.3.0
html5lib==0.999999999
ipython==5.0.0
ipython-genutils==0.1.0
lightdm-gtk-greeter-settings==1.2.1
louis==3.0.0
Markdown==2.4.1
menulibre==2.1.0
mugshot==0.3.1
packaging==16.7
pacman-mirrors==2.0
path.py==8.2.1
pexpect==4.2.1
pickleshare==0.7.3
Pillow==3.1.1
prompt-toolkit==1.0.3
psutil==4.3.1
psycopg2==2.6.2
ptyprocess==0.5.1
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1.3
pygobject==3.20.1
pyparsing==2.1.9
python-decouple==2.2
python-distutils-extra==2.39
pyxdg==0.25
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
team==1.0
traitlets==4.2.2
Unidecode==0.4.19
Unipath==1.0
urllib3==1.18
virtualenv==15.0.2
wcwidth==0.1.7
webencodings==0.5
